refactor: fix PHPStan argument type issues in Mastodon/Twitter API modules
- Cast substrings and string concatenations to int for media attachment
  ID lookups in Mastodon Media
- Pass explicit integer 1/0 to ThreadUser::setIgnored
- Replace boolean notify argument with Worker::PRIORITY_HIGH in
  Twitter Statuses Update
- Drop dead DBA::close call for the array returned by Contact::searchByName

+++ src/Module/Api/Mastodon/Media.php
@@ -94,7 +94,7 @@ class Media extends BaseApi
 		}
 
 		if (DI::mstdnAttachment()->isAttach($this->parameters['id']) && Attach::exists(['id' => substr((string) $this->parameters['id'], 7)])) {
-			$this->jsonExit(DI::mstdnAttachment()->createFromAttach(substr((string) $this->parameters['id'], 7)));
+			$this->jsonExit(DI::mstdnAttachment()->createFromAttach((int) substr((string) $this->parameters['id'], 7)));
 		}
 
 		$photo = Photo::selectFirst(['resource-id'], ['id' => $this->parameters['id'], 'uid' => $uid]);
@@ -112,7 +112,7 @@ class Media extends BaseApi
 			Post\Media::updateById(['description' => $request['description']], $this->parameters['id']);
 
 			try {
-				$attachment = DI::mstdnAttachment()->createFromId($this->parameters['id'] . '1');
+				$attachment = DI::mstdnAttachment()->createFromId((int) ($this->parameters['id'] . '1'));
 			} catch (InternalServerErrorException) {
 				$this->logAndJsonError(404, $this->errorFactory->RecordNotFound());
 			}
@@ -144,7 +144,7 @@ class Media extends BaseApi
 		}
 
 		if (DI::mstdnAttachment()->isAttach($id) && Attach::exists(['id' => substr((string) $id, 7)])) {
-			$this->jsonExit(DI::mstdnAttachment()->createFromAttach(substr((string) $id, 7)));
+			$this->jsonExit(DI::mstdnAttachment()->createFromAttach((int) substr((string) $id, 7)));
 		}
 
 		$this->logAndJsonError(404, $this->errorFactory->RecordNotFound());
